example/python/requirements.txt
Flask==1.1.2
google-cloud-kms==2.2.0
google-cloud-secret-manager==2.0.0
line-bot-sdk==1.17.0
gunicorn==20.0.4
